Use DirEntry.path for all os.path operations
In newer versions of Python 3.x, apparently DirEntry's will
automatically cast to a string. But in older versions this casting
doesn't happen automagically. Use d.path and f.path to explicitly
get the full filename path in string-form and pass that to any
downstream methods which take-in string-based filenames/paths.